intercase and intracase panel stubs
