+++ Core/src/org/sleuthkit/autopsy/commonfilesearch/Bundle.properties
@@ -1,6 +1,4 @@
 CommonFilesPanel.searchButton.text=Search
-CommonFilesPanel.withinDataSourceRadioButton.text=At least one match must appear in the data source selected  below:
-CommonFilesPanel.allDataSourcesRadioButton.text=Matches may be from any data source
 CommonFilesPanel.cancelButton.text=Cancel
 CommonFilesPanel.cancelButton.actionCommand=Cancel
 CommonFilesPanel.selectedFileCategoriesButton.text=Match on the following file categories:
@@ -17,5 +15,7 @@ CommonFilesPanel.jRadioButton1.text=jRadioButton1
 CommonFilesPanel.jRadioButton2.text=Correlate amongst external cases (compares files in current  case with Central Repo)
 CommonFilesPanel.intraCaseRadio.label=Correlate within current case only
 CommonFilesPanel.interCaseRadio.label=Correlate amongst all known cases (uses Central Repo)
-CommonFilesPanel.anCentralRepoCaseRadio.text_1=Matches may be from any Central Repo case
-CommonFilesPanel.specificCentralRepoCaseRadio.text_1=Matches must be from the following Central Repo case:
+IntraCasePanel.allDataSourcesRadioButton.text=Matches may be from any data source
+IntraCasePanel.withinDataSourceRadioButton.text=At least one match must appear in the data source selected  below:
+InterCasePanel.anCentralRepoCaseRadio.text=Matches may be from any Central Repo case
+InterCasePanel.specificCentralRepoCaseRadio.text=Matches must be from the following Central Repo case:
